20230907.0751 source update

+++ controller/Cboard.js
@@ -2,23 +2,35 @@ const {
     Board,
 } = require('../models');
 
-const boardList = (req,res)=>{
+const { stringToUuid } = require('./Cauth')
+
+// res.cookies(signedCookies) id값 복호화
+const getUserId = async (req) => {
+
+    if (req.signedCookies && req.signedCookies['logined'] && req.signedCookies['logined'].id) {
+        console.log('cookie', req.signedCookies['logined'].id);
+        return uuid = await stringToUuid(req.signedCookies['logined'].id);
+    } else {
+        return ''
+    }
+}
+
+const boardList = (req, res)=>{
     res.render('boardList');
 }
 
-const create_board = (req,res)=>{
+const create_board = (req, res)=>{
     res.render('postNew');
 }
 
-const edit_board = (req,res)=>{
+const edit_board = (req, res)=>{
     res.render('postEdit');
 }
 
-
-
-const user_id = '296b63ea-6f1c-4f18-9f10-382f4a80e1cd';
-
 const create_board_post = async (req,res)=>{
+
+    const user_id = await getUserId(req);
+
     const {title,content,event_time,bord_category} = req.body
     // user_id는 쿠키를 생성해서 req.cookies로 가져와야 될거 같긴 한데
     // 백앤드로 관계형 잘 설정되는지만 보려고 일단은 req에 같이 넣음
@@ -41,7 +53,11 @@ const create_board_post = async (req,res)=>{
     }
 }
 const edit_board_post = async(req,res)=>{
+
+    const user_id = await getUserId(req);
+
     const {id, title,content,event_time,bord_category} = req.body
+
     try{
         const board = await Board.update({
             title,
@@ -65,12 +81,17 @@ const edit_board_post = async(req,res)=>{
 }
 const delete_board = async (req, res) => {
 
+    const user_id = await getUserId(req);
+    if (user_id === '') return res.json({result:false});
+
     const { id } = req.body;
     
     try {
+
         const board = await Board.destroy({ where : { id }})
         if (board) {
             res.json({result:true});
+        
         } else {
             res.json({result:false});
         } 
@@ -82,12 +103,11 @@ const delete_board = async (req, res) => {
 
 }
 
-const boarduser_findone = async(req,res)=>{
+const boarduser_findone = async (req,res)=>{
 
     const {id} = req.body;
 
     try {
-        
         const board = await Board.findOne({
             attributes:['id', 'title', 'views', 'content', 'event_time', 'bord_category', 'createdAt'],
             where: {id}
@@ -105,15 +125,25 @@ const boarduser_findone = async(req,res)=>{
 }
 
 const boarduser_findall = async(req,res)=>{
-    const {id} = req.body;
-    console.log("id",id);
-    try{
+    
+    const user_id = await getUserId(req);
+   
+    try {
         const board = await Board.findAll({
             attributes:['id', 'title', 'views', 'content', 'event_time', 'bord_category', 'createdAt', 'poster_id'],
             limit:3
         })
+
         if(board){
-            res.json({result:true,board});
+            board.forEach(index => {
+                if (index.dataValues.poster_id === user_id) {
+                    index.dataValues.poster_check = true;
+                } else {
+                    index.dataValues.poster_check = false;
+                }
+            });
+            
+            res.json({result:true, board});
         }else{
             res.json({result:false});
         }
@@ -122,7 +152,6 @@ const boarduser_findall = async(req,res)=>{
         console.log(e);
     }
 }
-
 
 module.exports = {
     create_board,
